+++ Sources/Desktop/Resources/Locale/Righten_pl_PL.json
