+++ Sources/RightenWeb/Righten/.env
@@ -1,0 +1,12 @@
+FLASK_SECRET_KEY="Atr--h;s+&_|1y/M^azu8F=9lDlmI6+^5w,6P3-?O;+w_7Fwo$8CDmB=8.OW0Q1*ZS#N,l`ildn=N&E'CT\D3@5-_p\$uMNS$'9F5Uy;v7G~f1a'KP'lvTt6_q*f\.B"
+#FLASK_SQLALCHEMY_DATABASE_URI="sqlite:////righten/Database/Righten_mock.sqlite3"
+FLASK_SQLALCHEMY_DATABASE_URI='sqlite:///E:\\Projects\\Git\\Righten\\Sources\\Database\\Righten_mock.sqlite3'
+
+#Moved from .flaskenv, could be public
+FLASK_ENV=development
+FLASK_DEBUG=True #Works implicitly when FLASK_ENV=development
+FLASK_SQLALCHEMY_TRACK_MODIFICATIONS=True
+FLASK_PREFERRED_URL_SCHEME="https"
+#FLASK_SQLALCHEMY_ECHO=True
+FLASK_SQLALCHEMY_RECORD_QUERIES=True
+FLASK_LOG_FILENAME="/righten/righten.log"
